Move _get_torchao_version into its proper home
